remove(ng): move map settings to javascript folder
