+++ app/javascript/javascript/mapSettings.js
